+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## 0.0.1
+## 0.0.1 - Start package
 
-* TODO: Describe initial release.
+    * Start storybook.